+++ install_functions/manager.sh
@@ -29,7 +29,7 @@ function manager_startCluster() {
     lstart=$(grep -n "<cluster>" /var/wazuh-manager/etc/wazuh-manager.conf | cut -d : -f 1)
     lend=$(grep -n "</cluster>" /var/wazuh-manager/etc/wazuh-manager.conf | cut -d : -f 1)
 
-    eval 'sed -i -e "${lstart},${lend}s/<name>.*<\/name>/<name>wazuh_cluster<\/name>/" \
+    sed -i -e "${lstart},${lend}s/<name>.*<\/name>/<name>wazuh_cluster<\/name>/" \
         -e "${lstart},${lend}s/<node_name>.*<\/node_name>/<node_name>${winame}<\/node_name>/" \
         -e "${lstart},${lend}s/<node_type>.*<\/node_type>/<node_type>${manager_node_types[pos],,}<\/node_type>/" \
         -e "${lstart},${lend}s/<key>.*<\/key>/<key>${key}<\/key>/" \
@@ -38,7 +38,7 @@ function manager_startCluster() {
         -e "${lstart},${lend}s/<node>.*<\/node>/<node>${master_address}<\/node>/" \
         -e "${lstart},${lend}s/<hidden>.*<\/hidden>/<hidden>${hidden}<\/hidden>/" \
         -e "${lstart},${lend}s/<disabled>.*<\/disabled>/<disabled>${disabled}<\/disabled>/" \
-        /var/wazuh-manager/etc/wazuh-manager.conf'
+        /var/wazuh-manager/etc/wazuh-manager.conf
 
 }
 
@@ -63,7 +63,6 @@ function manager_configure(){
     common_logger -d "Setting provisional Wazuh indexer password."
     /var/wazuh-manager/bin/wazuh-manager-keystore -f indexer -k username -v admin
     /var/wazuh-manager/bin/wazuh-manager-keystore -f indexer -k password -v admin
-    common_logger "Wazuh manager vulnerability detection configuration finished."
 }
 
 function manager_install() {